requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
firebase-admin==7.1.0
pydantic-settings==2.10.1
python-dotenv==1.1.1
email-validator==2.3.0
python-jose==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
Pillow==11.3.0
